net: shell: Add suspend command
Move "suspend" command to a separate file.
