+++ subsys/net/lib/shell/CMakeLists.txt
@@ -25,6 +25,7 @@ zephyr_library_sources(resume.c)
 zephyr_library_sources(route.c)
 zephyr_library_sources(stacks.c)
 zephyr_library_sources(stats.c)
+zephyr_library_sources(suspend.c)
 
 zephyr_library_include_directories(
   ${ZEPHYR_BASE}/kernel/include
