+++ subsys/net/lib/shell/net_shell.c
@@ -1214,49 +1214,6 @@ usage:
 	return 0;
 }
 
-static int cmd_net_suspend(const struct shell *sh, size_t argc,
-			   char *argv[])
-{
-#if defined(CONFIG_NET_POWER_MANAGEMENT)
-	if (argv[1]) {
-		struct net_if *iface = NULL;
-		const struct device *dev;
-		int idx;
-		int ret;
-
-		idx = get_iface_idx(sh, argv[1]);
-		if (idx < 0) {
-			return -ENOEXEC;
-		}
-
-		iface = net_if_get_by_index(idx);
-		if (!iface) {
-			PR_WARNING("No such interface in index %d\n", idx);
-			return -ENOEXEC;
-		}
-
-		dev = net_if_get_device(iface);
-
-		ret = pm_device_action_run(dev, PM_DEVICE_ACTION_SUSPEND);
-		if (ret != 0) {
-			PR_INFO("Iface could not be suspended: ");
-
-			if (ret == -EBUSY) {
-				PR_INFO("device is busy\n");
-			} else if (ret == -EALREADY) {
-				PR_INFO("dehive is already suspended\n");
-			}
-		}
-	} else {
-		PR("Usage:\n");
-		PR("\tsuspend <iface index>\n");
-	}
-#else
-	PR_INFO("You need a network driver supporting Power Management.\n");
-#endif /* CONFIG_NET_POWER_MANAGEMENT */
-
-	return 0;
-}
 
 #if defined(CONFIG_WEBSOCKET_CLIENT)
 static void websocket_context_cb(struct websocket_context *context,
@@ -1365,8 +1322,6 @@ SHELL_STATIC_SUBCMD_SET_CREATE(net_cmd_udp,
 );
 
 SHELL_STATIC_SUBCMD_SET_CREATE(net_commands,
-	SHELL_CMD(suspend, NULL, "Suspend a network interface",
-		  cmd_net_suspend),
 	SHELL_CMD(tcp, &net_cmd_tcp, "Connect/send/close TCP connection.",
 		  cmd_net_tcp),
 	SHELL_CMD(udp, &net_cmd_udp, "Send/recv UDP packet", cmd_net_udp),
